include the PhpReferenceCompatibility into Zend\Session\Container, and remove the Zend/Session/Container/PhpReferenceCompatibility.php
