--- library/Zend/Session/Container/PhpReferenceCompatibility.php
@@ -1,37 +1,0 @@
-<?php
-/**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
- */
-
-namespace Zend\Session\Container;
-
-use Zend\Session\AbstractContainer;
-
-/**
- * Session storage container for PHP 5.3.4 and above.
- */
-abstract class PhpReferenceCompatibility extends AbstractContainer
-{
-    /**
-     * Retrieve a specific key in the container
-     *
-     * @param  string $key
-     * @return mixed
-     */
-    public function &offsetGet($key)
-    {
-        $ret = null;
-        if (!$this->offsetExists($key)) {
-            return $ret;
-        }
-        $storage = $this->getStorage();
-        $name    = $this->getName();
-        $ret =& $storage[$name][$key];
-
-        return $ret;
-    }
-}
